CLOSED - issue TAV-530: Default worker http://www.mygrid.org.uk/jira/browse/TAV-530
Added default worker at start-up if no workers present.
Also block non admins from creating or deleting workers (both client and server side).


git-svn-id: http://taverna.googlecode.com/svn@3244 bf327186-88b3-11dd-a302-d386e5130c1c

+++ trunk/taverna-service/taverna-rest/src/main/java/net/sf/taverna/service/rest/resources/WorkerResource.java
@@ -34,16 +34,21 @@ public class WorkerResource extends AbstractResource {
 
 	@Override
 	public void delete() {
-		if (worker!=null) {
-			if (!worker.isBusy()) {
-				deleteWorker();
+		if (getAuthUser().isAdmin()) {
+			if (worker!=null) {
+				if (!worker.isBusy()) {
+					deleteWorker();
+				}
+				else {
+					getResponse().setStatus(Status.CLIENT_ERROR_PRECONDITION_FAILED);
+				}
 			}
 			else {
-				getResponse().setStatus(Status.CLIENT_ERROR_PRECONDITION_FAILED);
+				getResponse().setStatus(Status.CLIENT_ERROR_NOT_FOUND);
 			}
 		}
 		else {
-			getResponse().setStatus(Status.CLIENT_ERROR_NOT_FOUND);
+			getResponse().setStatus(Status.CLIENT_ERROR_UNAUTHORIZED);
 		}
 	}
